+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Unreleased (v0.1.2)
+* Fail fast if ffmpeg cut samples are empty (< 1K).
+
 # v0.1.1
 * Add command to generate bash,fish & zsh completions `ab-av1 print-completions [SHELL]`.